OWIN: Change interop extension methods.
